iced_test

Function run

pub fn run(
    program: impl Program + 'static,
    tests_dir: impl AsRef<Path>,
) -> Result<(), Error>

Runs an Ice test suite for the given Program.

Any .ice tests will be parsed from the given directory and executed in an Emulator of the given Program.

Remember that an Emulator executes the real thing! Side effects will take place. It is up to you to ensure your tests have reproducible environments by leveraging [Preset][program::Preset].